+++ pyabc/predictor/predictor.py
@@ -1,6 +1,7 @@
 """Predictor implementations."""
 
 import numpy as np
+from scipy.stats import pearsonr
 from typing import Callable, List, Tuple, Union
 import copy
 import logging
@@ -62,6 +63,31 @@ class Predictor(ABC):
         """
 
 
+def wrap_fit_log(fit):
+    """Wrapper for fit logging."""
+    def wrapped_fun(self, x: np.ndarray, y: np.ndarray, w: np.ndarray):
+        start_time = time()
+
+        # actual fitting
+        ret = fit(self, x, y, w)
+
+        logger.info(f"Fitted {self} in {time() - start_time:.2f}s")
+        if self.log_pearson:
+            # shape: n_sample, n_out
+            y_pred = self.predict(x)
+            coeffs = [
+                pearsonr(y[:, i], y_pred[:, i])[0]
+                for i in range(y_pred.shape[1])
+            ]
+            logger.info(" ".join([
+                "Pearson correlations:",
+                *[f"{coeff:.3f}" for coeff in coeffs]]),
+            )
+
+        return ret
+    return wrapped_fun
+
+
 class SimplePredictor(Predictor):
     """Wrapper around generic predictor routines."""
 
@@ -72,6 +98,7 @@ class SimplePredictor(Predictor):
         normalize_labels: bool = True,
         joint: bool = True,
         weight_samples: bool = False,
+        log_pearson: bool = True,
     ):
         """
         Parameters
@@ -88,6 +115,8 @@ class SimplePredictor(Predictor):
         weight_samples:
             Whether to use importance sampling weights. Not that not all
             predictors may support weighted samples.
+        log_pearson:
+            Whether to log Pearson correlation coefficients after fitting.
         """
         self.predictor = predictor
         # only used if not joint
@@ -99,6 +128,8 @@ class SimplePredictor(Predictor):
         self.joint: bool = joint
         self.weight_samples: bool = weight_samples
 
+        self.log_pearson: bool = log_pearson
+
         # indices to use
         self.use_ixs: Union[np.ndarray, None] = None
 
@@ -108,6 +139,7 @@ class SimplePredictor(Predictor):
         self.mean_y: Union[np.ndarray, None] = None
         self.std_y: Union[np.ndarray, None] = None
 
+    @wrap_fit_log
     def fit(self, x: np.ndarray, y: np.ndarray, w: np.ndarray = None) -> None:
         """Fit the predictor to labeled data.
 
@@ -117,8 +149,6 @@ class SimplePredictor(Predictor):
         y: Targets, shape (n_sample, n_out).
         w: Weights, shape (n_sample,).
         """
-        start_time = time()
-
         # remove trivial features
         self.set_use_ixs(x=x)
         x = x[:, self.use_ixs]
@@ -154,8 +184,6 @@ class SimplePredictor(Predictor):
                 else:
                     predictor.fit(x, y_)
 
-        logger.info(f"Fitted {self} in {time() - start_time:.2f}s")
-
     def predict(self, x: np.ndarray, normalize: bool = False) -> np.ndarray:
         """Predict outputs using the model.
 
@@ -184,11 +212,16 @@ class SimplePredictor(Predictor):
         if self.joint:
             y = self.predictor.predict(x)
         else:
-            y = [predictor.predict(x) for predictor in self.single_predictors]
+            y = [predictor.predict(x).flatten()
+                 for predictor in self.single_predictors]
             y = np.array(y).T
 
         if not normalize and self.normalize_labels:
             y = y * self.std_y + self.mean_y
+
+        # some predictors may return flat arrays if n_out == 1
+        if y.ndim == 1:
+            y = y.reshape(-1, 1)
 
         return y
 
@@ -242,6 +275,7 @@ class LinearPredictor(SimplePredictor):
         normalize_labels: bool = True,
         joint: bool = True,
         weight_samples: bool = False,
+        log_pearson: bool = True,
         **kwargs,
     ):
         # check installation
@@ -262,6 +296,7 @@ class LinearPredictor(SimplePredictor):
             normalize_labels=normalize_labels,
             joint=joint,
             weight_samples=weight_samples,
+            log_pearson=log_pearson,
         )
 
     def fit(self, x: np.ndarray, y: np.ndarray, w: np.ndarray = None) -> None:
@@ -289,6 +324,7 @@ class LassoPredictor(SimplePredictor):
         normalize_features: bool = True,
         normalize_labels: bool = True,
         joint: bool = True,
+        log_pearson: bool = True,
         **kwargs,
     ):
         """Additional keyword arguments are passed on to the model."""
@@ -312,6 +348,7 @@ class LassoPredictor(SimplePredictor):
             normalize_labels=normalize_labels,
             joint=joint,
             weight_samples=False,
+            log_pearson=log_pearson,
         )
 
 
@@ -334,6 +371,7 @@ class GPPredictor(SimplePredictor):
         normalize_features: bool = True,
         normalize_labels: bool = True,
         joint: bool = True,
+        log_pearson: bool = True,
         **kwargs,
     ):
         """
@@ -361,6 +399,7 @@ class GPPredictor(SimplePredictor):
             normalize_labels=normalize_labels,
             joint=joint,
             weight_samples=False,
+            log_pearson=log_pearson,
         )
 
     def fit(self, x: np.ndarray, y: np.ndarray, w: np.ndarray = None) -> None:
@@ -457,12 +496,13 @@ class MLPPredictor(SimplePredictor):
     """
 
     def __init__(
-            self,
-            normalize_features: bool = True,
-            normalize_labels: bool = True,
-            joint: bool = True,
-            hidden_layer_sizes: Union[Tuple[int, ...], Callable] = None,
-            **kwargs,
+        self,
+        normalize_features: bool = True,
+        normalize_labels: bool = True,
+        joint: bool = True,
+        hidden_layer_sizes: Union[Tuple[int, ...], Callable] = None,
+        log_pearson: bool = True,
+        **kwargs,
     ):
         """Additional keyword arguments are passed on to the model.
 
@@ -497,6 +537,7 @@ class MLPPredictor(SimplePredictor):
             normalize_labels=normalize_labels,
             joint=joint,
             weight_samples=False,
+            log_pearson=log_pearson,
         )
 
     def fit(self, x: np.ndarray, y: np.ndarray, w: np.ndarray = None) -> None:
@@ -530,22 +571,25 @@ class HiddenLayerHandle:
 
     def __init__(
         self,
-        method: str = MAX,
+        method: Union[str, List[str]] = MEAN,
         n_layer: int = 1,
         max_size: int = np.inf,
-        alpha: float = 2.,
+        alpha: float = 1.,
     ):
         """
         Parameters
         ----------
         method:
-            Method to use.
+            Method to use. Can be any of:
 
             * "heuristic" bases the number of neurons on the number of samples
               to avoid overfitting. See
               https://stats.stackexchange.com/questions/181.
             * "mean" takes the mean of input and output dimension.
             * "max" takes the maximum of input and output dimension.
+
+            Additionally, a list of methods can be passed, in which case
+            the minimum over all is used.
         n_layer:
             Number of layers.
         max_size:
@@ -554,9 +598,12 @@ class HiddenLayerHandle:
             Factor used in "heuristic". The higher, the fewer neurons.
             Recommended is a value in the range 2-10.
         """
-        if method not in HiddenLayerHandle.METHODS:
-            raise ValueError(
-                f"Method {method} must be in {HiddenLayerHandle.METHODS}")
+        if isinstance(method, str):
+            method = [method]
+        for m in method:
+            if m not in HiddenLayerHandle.METHODS:
+                raise ValueError(
+                    f"Method {m} must be in {HiddenLayerHandle.METHODS}")
         self.method = method
         self.n_layer = n_layer
         self.max_size = max_size
@@ -576,17 +623,23 @@ class HiddenLayerHandle:
         -------
         hidden_layer_sizes: Tuple of hidden layer sizes.
         """
-        if self.method == HiddenLayerHandle.HEURISTIC:
-            # number of neurons
-            neurons = n_sample / (self.alpha * (n_in + n_out))
-            # divide by number of layers
-            neurons_per_layer = neurons / self.n_layer
-        elif self.method == HiddenLayerHandle.MEAN:
-            neurons_per_layer = 0.5 * (n_in + n_out)
-        elif self.method == HiddenLayerHandle.MAX:
-            neurons_per_layer = max(n_in, n_out)
-        else:
-            raise ValueError(f"Did not recognize method {self.method}.")
+        neurons_arr = []
+        for method in self.method:
+            if method == HiddenLayerHandle.HEURISTIC:
+                # number of neurons
+                neurons = n_sample / (self.alpha * (n_in + n_out))
+                # divide by number of layers
+                neurons /= self.n_layer
+            elif method == HiddenLayerHandle.MEAN:
+                neurons = 0.5 * (n_in + n_out)
+            elif method == HiddenLayerHandle.MAX:
+                neurons = max(n_in, n_out)
+            else:
+                raise ValueError(f"Did not recognize method {self.method}.")
+            neurons_arr.append(neurons)
+
+        # take minimum over proposed values
+        neurons_per_layer = min(neurons_arr)
 
         # cap
         neurons_per_layer = min(neurons_per_layer, self.max_size)